ai/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
pydantic>=2.0.0
websockets>=12.0
requests==2.31.0
gunicorn==21.2.0
